requirements.txt
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.0.3
nltk==3.4.1
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.1
Pillow==6.0.0
psycopg2==2.7.7
pyparsing==2.4.0
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
textblob==0.15.3
tweepy==3.7.0
urllib3==1.24.1
wordcloud==1.5.0
